Add a work-around to support async stride in NSSet.
The workaround in question is -getUnsafeObjects:count: which should
mirror the functionality in NSArray to place objects in a similar
array, unretained. This unfortunately still requires allocating more
than just an array to store the objects, but it's good enough. In the
long run, the allocations won't be what causes problems in most cases.