Replace binpack_t with slot_image_t.
In most cases, this produces far better results than binpack_t. The
downside is that it consumes a lot more memory and is slower, but that's
not too important for the situation.

In a few odd cases, binpack_t will do better, like when packing very
thin elements next to each other, but this usually results in packing
oddities that result in more font pages rather than fewer.